pyproject.toml
[tool.poetry]
name = "retail"
version = "0.1.0"
description = ""
authors = ["Craig Lenzen <lenzenc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
questionary = "^2.0.1"
swarm = {git = "https://github.com/openai/swarm.git"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
